requirements.txt
PyQt5~=5.15.10
openpyxl~=3.1.2
python-docx~=1.1.2